+++ app/routes/components/Input/Demo.tsx
@@ -2,8 +2,10 @@ import { Input } from "src/components/Input";
 
 export function Demo() {
   return (
-    <Input
-      placeholder="Type something..."
-    />
+    <div style={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
+      <Input size="sm" placeholder="Type something..." />
+      <Input placeholder="Type something..." />
+      <Input size="lg" placeholder="Type something..." />
+    </div>
   );
 }