+++ src/components/Input/CHANGELOG.md
@@ -2,3 +2,5 @@
 - 2025-09-13: Enhance Input component with more styling variables (
   `--input-font-size`, `--input-line-height`, `--input-caret-color`,
   `--input-placeholder-color`).
+- 2025-09-22: Support the `size` props (`sm`, `md`, `lg`) for Input component to
+  adjust height and font size. Etc.